Fixed critical section leak
DeleteCriticalSection() needed to be followed by delete(cs) -- the first call deallocates the mutex, the second deallocates the cs.

+++ Vortex-1.0/Vortex/Stream.h
@@ -31,14 +31,10 @@ public:
 	bool lockChunkTree;
 	BufferConfig(bool lock) {
 	    lockChunkTree = lock;
-	    if (lockChunkTree) 
-	        chunkCS = (CSType*)Syscall.MakeCS();
+	    if (lockChunkTree) chunkCS = (CSType*)Syscall.MakeCS();
 	}
 	~BufferConfig() {
-	    if (lockChunkTree) {
-		Syscall.DeleteCS(chunkCS);
-	        delete chunkCS;
-	    }
+	    if (lockChunkTree) Syscall.DeleteCS(chunkCS);
 	}
 };
